backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.24.3
python-multipart==0.0.6
werkzeug==3.0.1
python-dotenv==1.0.0